requirements.txt
flask 
flask-cors 
python-dotenv
PyPDF2 
langchain 
langchain_google_genai 
faiss-cpu 
huggingface_hub 
google-generativeai
